Fix 2.1 - Arreglar estilos y agregar pequeñas animaciones #166
