+++ frontend/src/assets/register/defaultRegisterImage.png
